requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
oauth2client>=4.1.3
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
openpyxl
python-docx
rapidfuzz 
Levenshtein
